re-organise EOF types
