+++ src/ethereum/prague/vm/instructions/__init__.py
@@ -14,7 +14,8 @@ implementations.
 """
 
 import enum
-from typing import Callable, Dict
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Callable, Dict
 
 from . import arithmetic as arithmetic_instructions
 from . import bitwise as bitwise_instructions
@@ -28,6 +29,19 @@ from . import memory as memory_instructions
 from . import stack as stack_instructions
 from . import storage as storage_instructions
 from . import system as system_instructions
+
+if TYPE_CHECKING:
+    from ..eof import EofVersion
+
+
+@dataclass
+class OpcodeStackItemCount:
+    """
+    Stack height count for an Opcode.
+    """
+
+    inputs: int
+    outputs: int
 
 
 class Ops(enum.Enum):
@@ -474,3 +488,202 @@ EOF1_TERMINATING_INSTRUCTIONS = (
     Ops.REVERT,
     Ops.INVALID,
 )
+
+
+op_stack_items: Dict[Ops, OpcodeStackItemCount] = {
+    Ops.STOP: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.ADD: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.MUL: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SUB: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.DIV: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SDIV: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.MOD: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SMOD: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.ADDMOD: OpcodeStackItemCount(inputs=3, outputs=1),
+    Ops.MULMOD: OpcodeStackItemCount(inputs=3, outputs=1),
+    Ops.EXP: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SIGNEXTEND: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.LT: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.GT: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SLT: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SGT: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.EQ: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.ISZERO: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.AND: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.OR: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.XOR: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.NOT: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.BYTE: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SHL: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SHR: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SAR: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.KECCAK: OpcodeStackItemCount(inputs=2, outputs=1),
+    Ops.SLOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.BLOCKHASH: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.COINBASE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.TIMESTAMP: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.NUMBER: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PREVRANDAO: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.GASLIMIT: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CHAINID: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.MLOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.MSTORE: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.MSTORE8: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.MSIZE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.MCOPY: OpcodeStackItemCount(inputs=3, outputs=0),
+    Ops.ADDRESS: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.BALANCE: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.ORIGIN: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CALLER: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CALLVALUE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CALLDATALOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.CALLDATASIZE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CALLDATACOPY: OpcodeStackItemCount(inputs=3, outputs=0),
+    Ops.CODESIZE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.CODECOPY: OpcodeStackItemCount(inputs=3, outputs=0),
+    Ops.GASPRICE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.EXTCODESIZE: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.EXTCODECOPY: OpcodeStackItemCount(inputs=4, outputs=0),
+    Ops.RETURNDATASIZE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.RETURNDATACOPY: OpcodeStackItemCount(inputs=3, outputs=0),
+    Ops.RETURNDATALOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.EXTCODEHASH: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.SELFBALANCE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.BASEFEE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.BLOBHASH: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.BLOBBASEFEE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.SSTORE: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.TLOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.TSTORE: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.JUMP: OpcodeStackItemCount(inputs=1, outputs=0),
+    Ops.JUMPI: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.PC: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.GAS: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.JUMPDEST: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.POP: OpcodeStackItemCount(inputs=1, outputs=0),
+    Ops.PUSH0: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH1: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH2: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH3: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH4: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH5: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH6: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH7: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH8: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH9: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH10: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH11: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH12: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH13: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH14: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH15: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH16: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH17: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH18: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH19: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH20: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH21: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH22: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH23: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH24: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH25: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH26: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH27: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH28: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH29: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH30: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH31: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.PUSH32: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP1: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP2: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP3: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP4: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP5: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP6: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP7: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP8: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP9: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP10: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP11: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP12: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP13: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP14: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP15: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DUP16: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.SWAP1: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP2: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP3: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP4: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP5: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP6: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP7: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP8: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP9: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP10: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP11: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP12: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP13: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP14: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP15: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.SWAP16: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.LOG0: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.LOG1: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.LOG2: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.LOG3: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.LOG4: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.DATALOAD: OpcodeStackItemCount(inputs=1, outputs=1),
+    Ops.DATALOADN: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DATASIZE: OpcodeStackItemCount(inputs=0, outputs=1),
+    Ops.DATACOPY: OpcodeStackItemCount(inputs=3, outputs=0),
+    Ops.RJUMP: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.RJUMPI: OpcodeStackItemCount(inputs=1, outputs=0),
+    Ops.RJUMPV: OpcodeStackItemCount(inputs=1, outputs=0),
+    Ops.CALLF: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.RETF: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.EOFCREATE: OpcodeStackItemCount(inputs=4, outputs=1),
+    Ops.RETURNCONTRACT: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.CREATE: OpcodeStackItemCount(inputs=3, outputs=1),
+    Ops.RETURN: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.CALL: OpcodeStackItemCount(inputs=7, outputs=1),
+    Ops.CALLCODE: OpcodeStackItemCount(inputs=7, outputs=1),
+    Ops.DELEGATECALL: OpcodeStackItemCount(inputs=6, outputs=1),
+    Ops.SELFDESTRUCT: OpcodeStackItemCount(inputs=1, outputs=0),
+    Ops.STATICCALL: OpcodeStackItemCount(inputs=6, outputs=1),
+    Ops.REVERT: OpcodeStackItemCount(inputs=2, outputs=0),
+    Ops.INVALID: OpcodeStackItemCount(inputs=0, outputs=0),
+    Ops.CREATE2: OpcodeStackItemCount(inputs=4, outputs=1),
+    Ops.EXTCALL: OpcodeStackItemCount(inputs=4, outputs=1),
+    Ops.EXTDELEGATECALL: OpcodeStackItemCount(inputs=3, outputs=1),
+    Ops.EXTSTATICCALL: OpcodeStackItemCount(inputs=3, outputs=1),
+}
+
+
+def map_int_to_op(opcode: int, eof_version: "EofVersion") -> Ops:
+    """
+    Get the opcode enum from the opcode value.
+
+    Parameters
+    ----------
+    opcode : `int`
+        The opcode value.
+    eof_version : `EofVersion`
+        The version of the EOF.
+
+    Returns
+    -------
+    opcode : `Ops`
+        The opcode enum.
+    """
+    from ..eof import EofVersion
+
+    try:
+        op = Ops(opcode)
+    except ValueError as e:
+        raise ValueError(f"Invalid opcode: {opcode}") from e
+
+    if eof_version == EofVersion.LEGACY and op in OPCODES_INVALID_IN_LEGACY:
+        raise ValueError(f"Invalid legacy opcode: {op}")
+    elif eof_version == EofVersion.EOF1 and op in OPCODES_INVALID_IN_EOF1:
+        raise ValueError(f"Invalid eof1 opcode: {op}")
+
+    return op
